Q3_Face_Feature_Localization/requirements.txt
opencv-python
mediapipe
Pillow
